Renamed to com.mattermost.rnbeta
