+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.mattermost"
+    package="com.mattermost.rnbeta"
     android:versionCode="1"
     android:versionName="1.0">
